refactor: extract api folder, useToggle and useFetch
