+++ src/components/__tests__/TodoForm.test.tsx
@@ -2,14 +2,16 @@ import { fireEvent, render, screen } from '@testing-library/react';
 import TodoForm from '../TodoForm';
 import '@testing-library/jest-dom';
 import userEvent from '@testing-library/user-event';
-import axios from 'axios';
+import { postTasks } from '../../api/api';
 
-jest.mock('axios');
+jest.mock('../../api/api');
+
 describe('TodoForm', () => {
   const mockTasks: any = {
     name: 'task name',
     completed: true,
   };
+
   it('should show input', () => {
     render(<TodoForm addTodo={mockTasks} />);
     expect(screen.getByLabelText('todo-input')).toBeInTheDocument();
@@ -23,7 +25,7 @@ describe('TodoForm', () => {
   });
 
   it('should clear input when click the button', () => {
-    (axios.post as jest.Mock).mockResolvedValue(mockTasks);
+    (postTasks as jest.Mock).mockResolvedValue(mockTasks);
     const input = render(<TodoForm addTodo={mockTasks} />).getByLabelText(
       'todo-input'
     );
@@ -50,10 +52,12 @@ describe('TodoForm', () => {
   });
 
   it('should execute a callback function with the input value as an argument when the buttion is pressed', () => {
-    (axios.post as jest.Mock).mockResolvedValue(mockTasks);
+    (postTasks as jest.Mock).mockImplementation(
+      async () => await Promise.resolve(mockTasks.name)
+    );
     render(<TodoForm addTodo={mockTasks} />);
     userEvent.type(screen.getByRole('textbox'), 'task 01');
     fireEvent.click(screen.getByText(/add items/i));
-    expect(axios.post).toBeCalledTimes(1);
+    expect(postTasks).toHaveBeenCalled();
   });
 });
